middleware theme_city created for THEME and City related routing in theme_city.js
